requirements.txt
pytz~=2023.3.post1
PyGithub~=2.1.1
python-dotenv~=1.0.0
allcities~=1.0.3
emoji-country-flag
pycountry